Move shared test setups to metabase-types
