+++ frontend/src/metabase-types/api/mocks/activity.ts
@@ -1,6 +1,7 @@
 import type {
   PopularItem,
   RecentCollectionItem,
+  RecentModel,
   RecentTableDatabaseInfo,
   RecentTableItem,
 } from "metabase-types/api";
@@ -41,6 +42,11 @@ export const createMockRecentCollectionItem = (
   },
   ...opts,
 });
+
+export const createMockRecentModel = (
+  model: Partial<RecentCollectionItem>,
+): RecentModel =>
+  createMockRecentCollectionItem({ ...model, model: "dataset" }) as RecentModel;
 
 export const createMockPopularTableItem = (
   opts?: Partial<RecentTableItem>,
