+++ frontend/src/metabase-types/api/mocks/search.ts
@@ -1,6 +1,10 @@
 import _ from "underscore";
 
-import type { SearchResponse, SearchResult } from "metabase-types/api";
+import type {
+  ModelResult,
+  SearchResponse,
+  SearchResult,
+} from "metabase-types/api";
 
 import { createMockCollection } from "./collection";
 
@@ -47,6 +51,11 @@ export const createMockSearchResult = (
     ...options,
   };
 };
+
+export const createMockModelResult = (
+  model: Partial<ModelResult> = {},
+): ModelResult =>
+  createMockSearchResult({ ...model, model: "dataset" }) as ModelResult;
 
 export const createMockSearchResults = ({
   items = [createMockSearchResult()],
