+++ frontend/src/metabase/nav/containers/MainNavbar/tests/setup.tsx
@@ -19,7 +19,6 @@ import {
   screen,
   waitForLoaderToBeRemoved,
 } from "__support__/ui";
-import type { ModelResult } from "metabase/browse/models";
 import { ROOT_COLLECTION } from "metabase/collections/constants";
 import type { DashboardState, StoreDashboard } from "metabase/redux/store";
 import {
@@ -28,7 +27,13 @@ import {
   createMockState,
 } from "metabase/redux/store/mocks";
 import * as iframeUtils from "metabase/utils/iframe";
-import type { Card, Dashboard, DashboardId, User } from "metabase-types/api";
+import type {
+  Card,
+  Dashboard,
+  DashboardId,
+  ModelResult,
+  User,
+} from "metabase-types/api";
 import {
   createMockCollection,
   createMockDatabase,
