+++ frontend/src/metabase/querying/notebook/components/NotebookDataPicker/EmbeddingDataPicker/tests/setup.tsx
@@ -4,12 +4,12 @@ import {
   setupSearchEndpoints,
 } from "__support__/server-mocks";
 import { renderWithProviders } from "__support__/ui";
-import { createMockModelResult } from "metabase/browse/models/test-utils";
 import type { EmbeddingEntityType } from "metabase/redux/store/embedding-data-picker";
 import { createMockState } from "metabase/redux/store/mocks";
 import { createMockEmbeddingDataPickerState } from "metabase/redux/store/mocks/embedding-data-picker";
 import type { Query } from "metabase-lib";
 import Question from "metabase-lib/v1/Question";
+import { createMockModelResult } from "metabase-types/api/mocks";
 import {
   createOrdersTable,
   createPeopleTable,
